Le degré est un float et pas un entier

+++ db/migrate/20151016091803_create_beers.rb
@@ -2,7 +2,7 @@ class CreateBeers < ActiveRecord::Migration
   def change
     create_table :beers do |t|
       t.string :name
-      t.integer :degree
+      t.float :degree
       t.text :description
       t.text :story
